Changes in safescale tenant get/set - safescale CLI get/set from/to user config file $HOME/.safescale/state.{json,yaml,toml} - Not needed anymore to get/set current tenant on backend side, every request MUST carry the concerned tenant

+++ lib/backend/resources/operations/network.go
@@ -311,7 +311,7 @@ func (instance *Network) Create(inctx context.Context, req abstract.NetworkReque
 
 		defer func() {
 			ferr = debug.InjectPlannedFail(ferr)
-			if ferr != nil && !req.KeepOnFailure {
+			if ferr != nil && !req.KeepOnFailure && !valid.IsNull(abstractNetwork) {
 				derr := svc.DeleteNetwork(context.Background(), abstractNetwork.ID)
 				derr = debug.InjectPlannedFail(derr)
 				if derr != nil {
